Improve testing of Ruby presence to prevent reinstallation.
Now, to check if a specific ruby with a specific version is installed in a
specific directory, we actually try to run the executable from that directory
and match the output version with the one specified by the user or expanded
by the installer.

Note that, in this case, if you provide `--no-reinstall` but specify a
different version or a different Ruby flavor, installation still takes place.
I'm not sure if this is a desirable effect, but for the moment let's just state
the fact that it can be done. Relaxing this condition should just be a matter of
removing a conditional expression.

To support this variation, related tests now installs a fake ruby executable in
the temporary install directory that output a baked version number. Again,
given the above consideration, the test suite for this feature might not be
complete, since there's no test against different version/flavor and same
installation directory.

Finally, I renamed all instances of `overriding` to `reinstall`.