Add switch to force the packager to be used.
I have dpkg tools installed on my Fedora machine to build Debian
packages. Using ruby-install fails here because I don't want to use
apt-get and it just so happens to be at the top of the branch
statement. This switch will allow us to override the default packager if necessary.