detect correct package manager on archlinux based distros
on archlinux based distros, one can have multiple package managers
installed. There are AUR pkgs available for homebrew, yum, apt and
so on.
This allows for someone in a archlinux box to build packages for others
distributions, or to simply install them when they don't conflict
with OS files, examples are packages that install into /opt/<pkg>

The previous detection method would look for apt, yum and would set
the packager command to match whatever it found first.
My archlinux box has yum, brew, however the correct package manager
is pacman. A quickfix is to search for pacman first and then the
others.
Usually EL or Debian based boxes do not have 'brew', 'pacman' or
other package managers installed or available as a package.

*Works on my machine*