Use apt-get instead of apt
When running non-interactively, this message appears in the logs:

    >>> Installing dependencies for ruby 3.2.1 ...
    WARNING: apt does not have a stable CLI interface. Use with caution in scripts.

This isn't a huge deal, but apt doesn't appear to respect some options
in /etc/apt/apt.conf.d/ when ran non-interactively. These options are
respected when running apt-get. This is at least true for
DPkg::Lock::Timeout, but may apply to other options as well.